Spell checking in pandocReferenceDefinition
This adds spell checking to the syntax region pandocReferenceDefinition
which is the region used for figure captions. This region does commonly
have a large chunk of text describing the figure, so having spell
checking is really helpful.

In adding spell checking to the definition region, spell checking for
the label and the address have been explicitly disabled.